Move solver monitor and friends to its own package.
