+++ ast/parser/Earthfile
@@ -1,7 +1,7 @@
 FROM alpine:3.15
 ARG USERARCH
 IF [ "$USERARCH" = "arm64" ]
-    FROM --platform=linux/arm64 arm64v8/openjdk:19-slim
+    FROM arm64v8/openjdk:19-slim
 ELSE
     FROM --platform=linux/amd64 openjdk:19-slim
 END
