+++ builder/builder_test.go
@@ -1,6 +1,7 @@
 package builder
 
 import (
+	"context"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -10,7 +11,7 @@ import (
 
 // TestTempEarthlyOutDir tests that tempEarthlyOutDir always returns the same directory
 func TestTempEarthlyOutDir(t *testing.T) {
-	b, _ := NewBuilder(nil, Opt{
+	b, _ := NewBuilder(context.Background(), Opt{
 		CleanCollection: cleanup.NewCollection(),
 	})
