+++ conslogging/conslogging_others.go
@@ -4,21 +4,10 @@
 package conslogging
 
 import (
+	"io"
 	"os"
-
-	"github.com/fatih/color"
 )
 
-// Current returns the current console.
-func Current(colorMode ColorMode, prefixPadding int, verbose bool) ConsoleLogger {
-	return ConsoleLogger{
-		consoleErrW:    os.Stderr,
-		errW:           os.Stderr,
-		colorMode:      colorMode,
-		saltColors:     make(map[string]*color.Color),
-		nextColorIndex: new(int),
-		prefixPadding:  prefixPadding,
-		mu:             &currentConsoleMutex,
-		verbose:        verbose,
-	}
+func getCompatibleStderr() io.Writer {
+	return os.Stderr
 }
