+++ conslogging/conslogging_windows.go
@@ -4,22 +4,12 @@
 package conslogging
 
 import (
+	"io"
 	"os"
 
-	"github.com/fatih/color"
 	"github.com/mattn/go-colorable"
 )
 
-// Current returns the current console.
-func Current(colorMode ColorMode, prefixPadding int, verbose bool) ConsoleLogger {
-	return ConsoleLogger{
-		consoleErrW:    colorable.NewColorable(os.Stderr),
-		errW:           colorable.NewColorable(os.Stderr),
-		colorMode:      colorMode,
-		saltColors:     make(map[string]*color.Color),
-		nextColorIndex: new(int),
-		prefixPadding:  prefixPadding,
-		mu:             &currentConsoleMutex,
-		verbose:        verbose,
-	}
+func getCompatibleStderr() io.Writer {
+	return colorable.NewColorable(os.Stderr)
 }
